+++ lib/src/domain/notifiers/simple_notifier.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:q_architecture/src/domain/mixins/simple_notifier_mixin.dart';
 
 abstract class SimpleNotifier<T> extends Notifier<T> with SimpleNotifierMixin {