+++ test/src/domain/notifiers/base_notifier_test.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: invalid_use_of_protected_member
 import 'package:either_dart/either.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:q_architecture/base_notifier.dart';
 import 'package:q_architecture/q_architecture.dart';